requirements.txt
python-dotenv==1.0.0
requests==2.31.0
mistralai==0.0.12 